Translate C++ code into Julia code.  Fix a bug.
Bug is in the C++ function `Place`.  The function will return a successful solution when `S` finally reach 81.  However, it will attempt to allocate to `LevelCount[81]` before that, which is out of bound as `LevelCount` is a length 81 array.  As C does not have bound check, this will usually cause unforeseen error if `LevelCount` happens to be allocated beside another critical variable.

As much as possible, the original coding and computational style is retained.
Otherwise, the original computational equivalent in Julia is achieved without change in intention.
Failing which, the original intention is understood and adapted in Julia.

Notably:-
- Index in Julia starts with 1 instead 0, unlike C++ (We can create types and override their `setindex!` and `getindex` but that defeats the purpose of this exercise).  Nevertheless, this actually make the Sudoku solver more intuitive as the original problem counts from 1 instead of 0.
- There is no direct equivalent to `printf`, `putc`, and `scanf` functions in Julia.  There is a `@printf` macro, which keeps much of the original meaning.  However `putc` and `scanf` have to be replaced.
- Julia may not play nice with global variables which are not declared as such in functions.