Fix bug in system test.
IO of any kind (text file IO in this case) are not easy to handle.  Decided to go with the slightly uglier string format, but at least a stable test could be done.

Programs working directly with STDIN/STDOUT are hard to test, so we will change it so that the output of the program will be the solution matrix instead of print out to screen.