Extract method `setEntry` from function `Place` and `SwapSeqEntries`.
There is much debate on how to optimize a program for speed.  For a usual program, it is advised not to optimize zealously.  A good habit is thus to keep methods short, modular, and reusable.  When a method gets too long and messy, extracting method that stands on its own is a good refactoring step.