Create function `clearEntry` and extract it from function `Place`.
Notice that we decided not to follow with the hack.  Hacks usually speedup at the expense of readability or natural algorithm flow.  They can be considered to speed up things in bottlenecks, but otherwise should be left alone in non-critical locations to keep the clarity.  A good gauge is the 80-20 rule.