Create type `gameState`, `Index`, 'Number`, and `LeftoverNumbers`.
This type is like a C's `struct`.  It will gather variable in a meaningful bundle.  Index, number, etc is clearer on intention than leaving it as `Int`.  It is also easier to pass a `struct` around functions, thus removing the need for the game state to be a global variable.

We also use type `Vector` and `Matrix` in this method, both of which are parametized types.  This allows use to describe a general structure that works for many underlying types.  For example, `Vector{Int}` and `Vector{Float64}` will work intuitively, as a vector of integer or vector of floats.  We will delve deeper into its power at a later time.