Write initialization of the game state in its own constructor.
The way to do object-oriented programming (OOP) is to encapsulate its own initialization inside its constructor.  This way, any creation of the object will ensure proper initialization.  A rule in programming is to write code that is difficult to use wrongly.  This habit of OOP avoid the case where a user forget to run initialization after object creation.

Sounds long-winded.  This point is that important: write code that is difficult to use wrongly.