Remove global variables in favour of passing a single object of `GameState`.
Most function now requires a `GameState` object to function properly (pun intended), however it is a small price to pay for getting rid of global.  This is usually the way to go for OOP languages, which encapsulate behaviour of a type as much as possible.  Some functions simplifies, by getting rid of `global` keywords.

A good way to visually check that we are done changing, is that there is no longer any ALL_CAPS variable except the constants `BLANK` and `ONES`.  This is once again the power of naming convention.

Ultimately, the fact that we have a system test that can be run and verify our output, and we have Git that can help us roll back to our previous code, we are able to edit code confidently.