Create type `ComponentState` to store state of either blocks, rows, or columns.
Once again, encapsulation allows us to hide ugly subscriptions like `g.leftoverNumbers_block[gs.block_of[square]]` into `gs.blocks[square]` and `s.data[s.location[index]]`, without loss of meaning.  Julia is very powerful in the sense that it allows us to overwrite the accessor `[]` by giving our own implementation of `getindex` and `setindex!`.

A point to note (but we have not enforced), is that in Julia, methods with side effect has `!` in their name.  This is to alert user that a side-effect exists, so feed inputs with care.