Define types `Cell`, `CellsState`, `SequenceState`, and `LevelCount`.
We try to perform major encapsulation.  Construction of each type is isolated inside its constructor.  This make the code easier to find and understand.  We observe that `SolveSudoku.jl` continues to get clearer in meaning, at the expense of extra helper functions defined in `operations.jl`.

Observe the abundant use of Julia constructs in this commits.  Each of them make the code cleaner and more concise (occasionally at a slight expense of efficiency).

Nevertheless, we change indexing of cell from 1, 2,... 81 to (1, 1),.., (1, 9), (2, 1), ..., (9, 9).  So there is some expected slowdown in the program, once again for the sake of clarity.