Overhaul the IO format.
Major changes.  Remove console input.  Replace with initialization with an input matrix of the state.  Output is the set of solutions found.  For our test run, there is only one solution.

We also fix the bug that `LevelCount` needs length 82.  Now it only uses 81.  COUNT should have been 17697 and Total COUNT is 24394.