Clean up the code files.
Comments are mostly in, which is helpful to any one reading the code any time.  The naming shifted from program language (describing what C/C++ does) to domain language (describing overview on solving Sudoku).  The machineries are hidden in the file `operation.jl`.  They are simple functions that are easy to understand when reading by itself, and act as bridge between domain language and the underlying program language.

There is much that can be improved, but by the 80-20 rule, this is a good point to stop =)